minor refactors + example of styled component setup
