+++ src/styled/BodyText.ts
@@ -1,0 +1,8 @@
+import { css } from 'styled-components';
+
+const BodyText = css`
+    font-size: var(--body-size);
+    line-height: var(--body-lh);
+`;
+
+export default BodyText;
