+++ src/styled/index.ts
@@ -1,0 +1,1 @@
+export { default as BodyText } from './BodyText';
